+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Exposed `lookupObjectMarker(value, type?)` and `lookupObjectMarkerAll(value)` methods to prepare handler
 - Added `Widget#queryFn(query)` method
+- Fixed symbol stringifing in `struct` view
 
 ## 1.0.0-beta.41 (06-10-2020)